refactor(workflows): move Archive/Unarchive from the header into the Settings panel
