+++ apps/web/modules/workflows/components/inspector/workflow-settings-section.tsx
@@ -1,11 +1,16 @@
 "use client";
 
 import { useAtomValue, useSetAtom } from "jotai";
+import { ArchiveIcon, ArchiveRestoreIcon } from "lucide-react";
+import { useState } from "react";
 import { useTranslation } from "react-i18next";
 import { cn } from "@/lib/cn";
+import { Button } from "@/modules/ui/components/button";
+import { ConfirmationModal } from "@/modules/ui/components/confirmation-modal";
 import { Input } from "@/modules/ui/components/input";
 import { Label } from "@/modules/ui/components/label";
 import { InspectorSection } from "@/modules/workflows/components/inspector/workflow-inspector-section";
+import { useWorkflowBuilder } from "@/modules/workflows/hooks/use-workflow-builder";
 import {
   setWorkflowDescriptionAtom,
   setWorkflowNameAtom,
@@ -14,15 +19,28 @@ import {
 } from "@/modules/workflows/state/editor";
 
 interface SettingsSectionProps {
+  workflowId: string;
+  isReadOnly: boolean;
   canEditMetadata: boolean;
 }
 
-export const SettingsSection = ({ canEditMetadata }: Readonly<SettingsSectionProps>) => {
+export const SettingsSection = ({
+  workflowId,
+  isReadOnly,
+  canEditMetadata,
+}: Readonly<SettingsSectionProps>) => {
   const { t } = useTranslation();
   const workflowName = useAtomValue(workflowNameAtom);
   const workflowDescription = useAtomValue(workflowDescriptionAtom);
   const setWorkflowName = useSetAtom(setWorkflowNameAtom);
   const setWorkflowDescription = useSetAtom(setWorkflowDescriptionAtom);
+  const builder = useWorkflowBuilder({ workflowId, isReadOnly, loadOnMount: false });
+  const [isArchiveModalOpen, setIsArchiveModalOpen] = useState(false);
+
+  const handleArchiveConfirm = async () => {
+    await builder.archive();
+    setIsArchiveModalOpen(false);
+  };
 
   return (
     <InspectorSection title={t("workspace.workflows.settings_title")} defaultOpen>
@@ -48,13 +66,57 @@ export const SettingsSection = ({ canEditMetadata }: Readonly<SettingsSectionPro
             disabled={!canEditMetadata}
             rows={3}
             className={cn(
-              "flex w-full rounded-md border border-slate-300 bg-white px-3 py-2 text-sm text-slate-800 placeholder:text-slate-400 focus:border-brand-dark focus:outline-none focus:ring-2 focus:ring-slate-400 focus:ring-offset-2 disabled:cursor-not-allowed disabled:opacity-50 dark:border-slate-500 dark:text-slate-300"
+              "flex w-full rounded-md border border-slate-300 bg-white px-3 py-2 text-sm text-slate-800 placeholder:text-slate-400 focus:border-brand-dark focus:ring-2 focus:ring-slate-400 focus:ring-offset-2 focus:outline-none disabled:cursor-not-allowed disabled:opacity-50 dark:border-slate-500 dark:text-slate-300"
             )}
             placeholder={t("workspace.workflows.workflow_description_placeholder")}
             onChange={(event) => setWorkflowDescription(event.target.value)}
           />
         </div>
+
+        <div className="flex flex-col gap-2 border-t border-slate-200 pt-4">
+          <Label>{t("workspace.workflows.archive_workflow")}</Label>
+          <p className="text-xs text-slate-500">{t("workspace.workflows.archive_workflow_description")}</p>
+          {builder.isArchived ? (
+            <Button
+              type="button"
+              variant="secondary"
+              size="sm"
+              className="w-fit"
+              onClick={builder.unarchive}
+              loading={builder.isTransitioning}
+              disabled={isReadOnly || builder.isTransitioning}>
+              {t("common.unarchive")}
+              <ArchiveRestoreIcon />
+            </Button>
+          ) : (
+            // The confirmation modal owns the spinner, so this button only disables while
+            // another lifecycle action runs.
+            <Button
+              type="button"
+              variant="secondary"
+              size="sm"
+              className="w-fit"
+              onClick={() => setIsArchiveModalOpen(true)}
+              disabled={isReadOnly || builder.isTransitioning || builder.isSaving}>
+              {t("common.archive")}
+              <ArchiveIcon />
+            </Button>
+          )}
+        </div>
       </div>
+
+      <ConfirmationModal
+        open={isArchiveModalOpen}
+        setOpen={setIsArchiveModalOpen}
+        title={t("workspace.workflows.archive_confirm_title")}
+        body={t("workspace.workflows.archive_confirm_body")}
+        buttonText={t("common.archive")}
+        buttonVariant="destructive"
+        buttonLoading={builder.isTransitioning}
+        isButtonDisabled={isReadOnly || builder.isTransitioning}
+        onConfirm={handleArchiveConfirm}
+        Icon={ArchiveIcon}
+      />
     </InspectorSection>
   );
 };
