+++ apps/web/modules/workflows/components/workflow-header-cta.tsx
@@ -1,12 +1,10 @@
 "use client";
 
 import { useAtomValue, useSetAtom } from "jotai";
-import { ArchiveIcon, ArchiveRestoreIcon, Loader2Icon, SaveIcon } from "lucide-react";
+import { Loader2Icon, SaveIcon } from "lucide-react";
 import { useSelectedLayoutSegment } from "next/navigation";
-import { useState } from "react";
 import { useTranslation } from "react-i18next";
 import { Button } from "@/modules/ui/components/button";
-import { ConfirmationModal } from "@/modules/ui/components/confirmation-modal";
 import { Switch } from "@/modules/ui/components/switch";
 import { useWorkflowBuilder } from "@/modules/workflows/hooks/use-workflow-builder";
 import { isCanvasLockedAtom, workflowAtom } from "@/modules/workflows/state/editor";
@@ -22,7 +20,6 @@ export const WorkflowHeaderCta = ({ workflowId, isReadOnly }: Readonly<WorkflowH
   const workflow = useAtomValue(workflowAtom);
   const setCanvasLocked = useSetAtom(isCanvasLockedAtom);
   const builder = useWorkflowBuilder({ workflowId, isReadOnly, loadOnMount: false });
-  const [isArchiveModalOpen, setIsArchiveModalOpen] = useState(false);
 
   // Header Save is the "end this editing session" action — persist, then re-lock the canvas so
   // the user explicitly opts back into edit mode. Per-node modal Save just persists.
@@ -39,11 +36,6 @@ export const WorkflowHeaderCta = ({ workflowId, isReadOnly }: Readonly<WorkflowH
     }
   };
 
-  const handleArchiveConfirm = async () => {
-    await builder.archive();
-    setIsArchiveModalOpen(false);
-  };
-
   // Only the edit tab gets the lifecycle controls — the runs tab is read-only.
   if (segment !== null) return null;
   if (!workflow) return null;
@@ -53,77 +45,36 @@ export const WorkflowHeaderCta = ({ workflowId, isReadOnly }: Readonly<WorkflowH
 
   return (
     <div className="flex items-center gap-3">
-      {isArchived ? (
-        <Button
-          type="button"
-          variant="secondary"
-          size="sm"
-          onClick={builder.unarchive}
-          loading={builder.isTransitioning}
-          disabled={isReadOnly || builder.isTransitioning}>
-          {t("common.unarchive")}
-          <ArchiveRestoreIcon />
-        </Button>
-      ) : (
-        <>
-          {/* Archive first — opens a confirmation modal before the destructive transition. The
-              spinner lives on the modal's confirm button, so this one only disables (no spinner)
-              while another lifecycle action runs. */}
-          <Button
-            type="button"
-            variant="secondary"
-            size="sm"
-            onClick={() => setIsArchiveModalOpen(true)}
-            disabled={isReadOnly || builder.isTransitioning || builder.isSaving}>
-            {t("common.archive")}
-            <ArchiveIcon />
-          </Button>
-          {/* Save sits second as a secondary action — where users instinctively look for it. The
-              trailing Save icon turns into the spinner while saving (single icon slot, never both),
-              so the label never shifts. */}
-          <Button
-            type="button"
-            variant="secondary"
-            size="sm"
-            onClick={handleSave}
-            disabled={!builder.canEditMetadata || builder.isTransitioning || builder.isSaving}
-            className="min-w-[6rem] justify-center">
-            {t("common.save")}
-            {builder.isSaving ? <Loader2Icon className="animate-spin" /> : <SaveIcon />}
-          </Button>
-          {/* Lifecycle control as a toggle: the switch position communicates the current state
-              (on = running) instead of an Enable/Disable action label the user has to invert.
-              The shell matches the sibling h-8 buttons so it reads as part of the button row. */}
-          <label
-            htmlFor="workflow-enabled-toggle"
-            className="flex h-8 cursor-pointer items-center gap-2 rounded-md border border-primary/5 bg-secondary px-3 text-xs font-medium text-secondary-foreground has-[button:disabled]:cursor-not-allowed has-[button:disabled]:opacity-50">
-            <Switch
-              id="workflow-enabled-toggle"
-              checked={isActive}
-              disabled={isReadOnly || builder.isTransitioning || builder.isSaving}
-              onCheckedChange={handleActiveChange}
-            />
-            {builder.isTransitioning ? (
-              <Loader2Icon className="size-4 animate-spin" aria-hidden="true" />
-            ) : (
-              <span>{isActive ? t("common.enabled") : t("common.disabled")}</span>
-            )}
-          </label>
-        </>
-      )}
-
-      <ConfirmationModal
-        open={isArchiveModalOpen}
-        setOpen={setIsArchiveModalOpen}
-        title={t("workspace.workflows.archive_confirm_title")}
-        body={t("workspace.workflows.archive_confirm_body")}
-        buttonText={t("common.archive")}
-        buttonVariant="destructive"
-        buttonLoading={builder.isTransitioning}
-        isButtonDisabled={isReadOnly || builder.isTransitioning}
-        onConfirm={handleArchiveConfirm}
-        Icon={ArchiveIcon}
-      />
+      {/* The trailing Save icon turns into the spinner while saving (single icon slot, never
+          both), so the label never shifts. Archive/Unarchive live in the Settings panel. */}
+      <Button
+        type="button"
+        variant="secondary"
+        size="sm"
+        onClick={handleSave}
+        disabled={!builder.canEditMetadata || builder.isTransitioning || builder.isSaving}
+        className="min-w-[6rem] justify-center">
+        {t("common.save")}
+        {builder.isSaving ? <Loader2Icon className="animate-spin" /> : <SaveIcon />}
+      </Button>
+      {/* Lifecycle control as a toggle: the switch position communicates the current state
+          (on = running) instead of an Enable/Disable action label the user has to invert.
+          The shell matches the sibling h-8 buttons so it reads as part of the button row. */}
+      <label
+        htmlFor="workflow-enabled-toggle"
+        className="flex h-8 cursor-pointer items-center gap-2 rounded-md border border-primary/5 bg-secondary px-3 text-xs font-medium text-secondary-foreground has-[button:disabled]:cursor-not-allowed has-[button:disabled]:opacity-50">
+        <Switch
+          id="workflow-enabled-toggle"
+          checked={isActive}
+          disabled={isReadOnly || isArchived || builder.isTransitioning || builder.isSaving}
+          onCheckedChange={handleActiveChange}
+        />
+        {builder.isTransitioning ? (
+          <Loader2Icon className="size-4 animate-spin" aria-hidden="true" />
+        ) : (
+          <span>{isActive ? t("common.enabled") : t("common.disabled")}</span>
+        )}
+      </label>
     </div>
   );
 };
